Text extractor microservice modified.
